flatten the digit groups to an int

+++ gearshift/src/main/kotlin/org/sugr/gearshift/viewmodel/ext/readability.kt
@@ -8,11 +8,11 @@ fun Long.readableFileSize() : String {
     if (this <= 0) return "0 B";
 
     val units = arrayOf("B", "KB", "MB", "GB", "TB")
-    val digitGroups = (Math.log10(this.toDouble())/Math.log10(1024.0));
-    val scaledSize = this / Math.pow(1024.0, digitGroups);
+    val digitGroups = (Math.log10(this.toDouble())/Math.log10(1024.0)).toInt();
+    val scaledSize = this / Math.pow(1024.0, digitGroups.toDouble());
     val format = if (scaledSize < 100) "#,##0.##" else "#,##0.#"
 
-    return DecimalFormat(format).format(scaledSize) + " " + units[digitGroups.toInt()]
+    return DecimalFormat(format).format(scaledSize) + " " + units[digitGroups]
 }
 
 fun Long.readableRemainingTime(ctx : Context) : String {
